Feature/auth

+++ pages/api/deleteSnippet.js
@@ -1,9 +1,19 @@
-import { deleteSnippet } from "../../utils/Fauna";
-export default async function handler(req, res) {
+import { deleteSnippet, getSnippetById } from "../../utils/Fauna";
+import { withApiAuthRequired, getSession } from "@auth0/nextjs-auth0";
+
+export default withApiAuthRequired(async function handler(req, res) {
+  const session = getSession(req, res);
+  const userId = session.user.sub;
+
   if (req.method !== "DELETE") {
     return res.status(405).json({ msg: "Method not allowed" });
   }
   const { id } = req.body;
+  const dbRecord = await getSnippetById(id);
+
+  if (!dbRecord || dbRecord.data.userId !== userId) {
+    return res.status(401).json({ msg: "Cannot delete" });
+  }
 
   try {
     const deletedSnippet = await deleteSnippet(id);
@@ -12,4 +22,4 @@ export default async function handler(req, res) {
     console.error(err);
     res.status(500).json({ msg: "Something went wrong." });
   }
-}
+});
